Move standard and common logical type implementations into separate module
